+++ packages/gennai_local_runner/src/gennai_local_runner/app.py
@@ -16,7 +16,7 @@ except Exception:  # pragma: no cover
     HTMLResponse = None
     JSONResponse = None
 
-APP_VERSION = "0.5.0"
+APP_VERSION = "0.6.0"
 
 
 HTML = r"""